Incremental patch for enhancement of program with command line arguments #25 /* The current program: Creates new namespace whenever a new combination of options comes in. Dedup and restore files. List namespaces and files that are deduped using perticular namespace. Displays the information of all the namespaces and the information of perticular namespace. Delete, edit and reset namespace. Delete stub and reset catalog. Provides information to use commands of YADL.
.yadl hidden file yet to be created.
Variables need to be update with meaningful names.
Used some static size yet to be modified.
*/

+++ src/namespace.c
@@ -45,6 +45,9 @@ print_usage (FILE *stream, int exit_code)
                 "$> yadl --info/-i -n all\n"
                 "\nList all namespace:\n"
                 "$> yadl --list/-l\n"
+                "\nEdit namespace:\n"
+                "$> yadl --edit/-e -n <namespace_name>"
+                        " [--desc <namespace_description>]\n"
                 "\nReset namespace:\n"
                 "$> yadl --reset/-R -n <namespace_name>\n"
                 "\nDelete namespace:\n"
@@ -76,6 +79,7 @@ create_namespace(char *namespace_path)
         char    content[1024]   =       "";
         char    path[1024]      =       "";
         DIR     *dp             =       NULL;
+        DIR     *store_dp       =       NULL;
         struct stat st;
 
         if (namespace_path == NULL || nm.store_path == NULL ||
@@ -155,6 +159,17 @@ create_namespace(char *namespace_path)
                 if (ret < 0)
                         goto out;
                 printf("namespace created %s\n", filename);
+                memset(path, 0, 1024);
+                sprintf(path, "%s/store_block", nm.store_path);
+                store_dp = opendir(path);
+                if (NULL == store_dp) {
+                        ret = mkdir(path, 0777);
+                        if (ret < 0) {
+                                fprintf(stderr, "%s\n", strerror(errno));
+                                goto out;
+                        }
+                }
+                printf("store created %s\n", path);
         } else {
                 printf("namespace already present\n");
         }
@@ -290,6 +305,63 @@ out:
         return ret;
 }
 
+int
+edit_namespace(char *namespace_path)
+{
+        int     ret             =       -1;
+        int     fd              =       -1;
+        char    filename[1024]  =       "";
+        char    content[1024]   =       "";
+        char    replace[1024]   =       "";
+        char    buffer[1024]    =       "";
+        char    *p              =       NULL;
+
+        if (nm.namespace_name == NULL || nm.desc == NULL ||
+                namespace_path == NULL) {
+                goto out;
+        }
+        sprintf(replace, "desc:%s", nm.desc);
+        sprintf(filename, "%s/%s.yadl", namespace_path,
+                nm.namespace_name);
+        printf("%s :\n", filename);
+        fd = open(filename, O_RDONLY, S_IRUSR|S_IWUSR);
+        if (fd < 1) {
+                fprintf(stderr, "%s\n", strerror(errno));
+                goto out;
+        }
+        ret = read(fd, content, 1024);
+        if (ret < 0) {
+                fprintf(stderr, "%s\n", strerror(errno));
+                goto out;
+        }
+        p = strstr(content, "desc:");
+        if (!p)
+                goto out;
+
+        strncpy(buffer, content, p-content);
+        buffer[p-content] = '\0';
+        sprintf(buffer+(p-content), "%s", replace);
+        printf("%s\n", buffer);
+        if (fd != -1)
+                close(fd);
+        fd = open(filename, O_WRONLY | O_TRUNC, S_IRUSR|S_IWUSR);
+        if (fd < 1) {
+                fprintf(stderr, "%s\n", strerror(errno));
+                goto out;
+        }
+        if (-1 == write(fd, buffer, strlen(buffer))) {
+                fprintf(stderr, "%s\n", strerror(errno));
+                goto out;
+        }
+        ret = 0;
+out:
+        if (fd != -1)
+                close(fd);
+        if (ret < 0)
+                print_usage(stderr, 1);
+        return ret;
+}
+
 /*Function to delete the namespace.
 Input:
         char *namespace_path : Path of the namespace.
@@ -322,7 +394,7 @@ delete_namespace(char *namespace_path)
                                 if (strcmp(dir->d_name, namespace_file) == 0) {
                                         sprintf(filename, "%s/%s",
                                                 namespace_path, namespace_file);
-                                        ret = file_operation(reset, NULL,
+                                        ret = file_operation(reset, "delete",
                                                 namespace_path);
                                         if (ret == -1 || ret == 1) {
                                                 goto out;
@@ -423,6 +495,7 @@ file_operation(enum FLAG flag, char *filename, char *namespace_path)
         char    *saveptr1       =       NULL;
         char    *saveptr2       =       NULL;
         char    path[1024]      =       "";
+        char    confirm         =       -1;
 
         if (namespace_path == NULL) {
                 goto out;
@@ -528,15 +601,22 @@ file_operation(enum FLAG flag, char *filename, char *namespace_path)
                         goto out;
         }
         if (flag == delete_file) {
-                if (filename == NULL) {
-                        printf("File name requried :"
-                        "Try $>yadl --help for more information\n");
+                printf("Do you want to continue with deletion?[y/n]");
+                if (scanf("%c", &confirm) <= 0) {
+                        fprintf(stderr, "%s\n", strerror(errno));
                         goto out;
                 }
-                ret = delete_stub_store(nm1.store_path, filename);
-                if (ret < 0)
-                        goto out;
-                printf("\nStub deleted\n");
+                if (confirm == 'y' || confirm == 'Y') {
+                        if (filename == NULL) {
+                                printf("File name requried :"
+                                "Try $>yadl --help for more information\n");
+                                goto out;
+                        }
+                        ret = delete_stub_store(nm1.store_path, filename);
+                        if (ret < 0)
+                                goto out;
+                        printf("\nStub deleted\n");
+                }
         }
         if (flag == list) {
                 ret = readfilecatalog();
@@ -552,6 +632,8 @@ file_operation(enum FLAG flag, char *filename, char *namespace_path)
                         ret = 1;
                         goto out;
                 }
+                if (strcmp(filename, "delete") == 0)
+                        goto out;
                 ret = mkdir(path, 0777);
                 if (ret < 0) {
                         fprintf(stderr, "%s\n", strerror(errno));
@@ -595,7 +677,7 @@ start_program(int argc, char **argv, char *namespace_path)
         enum    FLAG flag               =       -1;
         int     option_index            =        0;
 
-        const char *short_options = "cn:p:h:s:df:ilRr";
+        const char *short_options = "cn:p:h:s:df:ilRre";
 
         const struct option long_options[] = {
                 {"create",          no_argument,            0,   'c'},
@@ -612,7 +694,8 @@ start_program(int argc, char **argv, char *namespace_path)
                 {"delete",          no_argument,            0,   'd'},
                 {"info",            no_argument,            0,   'i'},
                 {"list",            no_argument,            0,   'l'},
-                {"reset",            no_argument,            0,   'R'},
+                {"reset",           no_argument,            0,   'R'},
+                {"edit",            no_argument,            0,   'e'},
                 {"help",            no_argument,            0,   0},
                 {0,                 0,                      0,   0 }
         };
@@ -733,6 +816,10 @@ start_program(int argc, char **argv, char *namespace_path)
                         flag = reset;
                         break;
 
+                case 'e':
+                        flag = edit;
+                        break;
+
                 case '?':
                         print_usage (stderr, 1);
                         goto out;
@@ -757,7 +844,7 @@ start_program(int argc, char **argv, char *namespace_path)
                         goto out;
                 break;
         case 1:
-                ret = create_namespace(namespace_path);
+                ret = edit_namespace(namespace_path);
                 if (ret == -1)
                         goto out;
                 break;
